Add play timers
A play timer is just a generic wrapper for a tracker. Timers can be
compared and should receive ticks. The tick method will increment an
internal counter which are for pseudo-minutes/seconds. If the tick is
received more often, time will move faster and vice versa.

RWGameTimer just provides a singleton, timer collection, and a few
helpful methods used for display. A GameTimer, when it receives a tick,
will tick the timer for the current player. This is tighter coupling
than I normally prefer, but in this instance it's OK.