Fix game over screen not displaying
This addresses the issue of determining the winner of the overall game.
The winner method had to be called at the end of each turn and it
wasn't. This was something fixed within the `playPosition:withPlayer`
method for each board, but wasn't applied to the RWGame.

Calling the method causes the game to calculate the current winner and
emit a notification if/when there is a winner or there is a stalemate
(which is just another type of winner).